replaced evaluate button with automatic evaluation on every line click
